After implementing support for _building_ embargoed Git versions, now it
is time to add support for releasing them, too. This is a two-stage
process: first prepare pseudo-releases in the embargoed `git`
repository, and then, on the day the embargo is lifted, recreate the
corresponding GitHub Releases in the `git-for-windows/git` repository.

+++ repository-updates.js
@@ -29,28 +29,30 @@ const getWorkflowRunArtifact = async (context, token, owner, repo, workflowRunId
 }
 
 const mergeBundle = (gitDir, worktree, bundlePath, refName) => {
-  callGit(['--git-dir', gitDir, 'fetch', bundlePath, refName])
+  callGit(['--git-dir', gitDir, 'fetch', bundlePath, refName.startsWith('refs/tags/') ? `${refName}:${refName}` : refName])
   // If there is nothing to merge, return early
-  if (callGit(['--git-dir', gitDir, 'rev-list', '--count', `${refName}..FETCH_HEAD`]) === '0') return
+  if (callGit(['--git-dir', gitDir, 'rev-list', '--count', 'HEAD..FETCH_HEAD']) === '0') return
 
   if (worktree) {
     callGit(['branch', '-M', refName], worktree)
     callGit(['merge', '--no-edit', 'FETCH_HEAD'], worktree)
   } else {
     // If it fast-forwards, we do not need a worktree
-    if (callGit(['--git-dir', gitDir, 'rev-list', '--count', `FETCH_HEAD..${refName}`]) === '0') {
-      callGit(['--git-dir', gitDir, 'update-ref', `refs/heads/${refName}`, 'FETCH_HEAD'])
+    if (callGit(['--git-dir', gitDir, 'rev-list', '--count', `FETCH_HEAD..HEAD`]) === '0') {
+      callGit(['--git-dir', gitDir, 'update-ref', `HEAD`, 'FETCH_HEAD^{commit}'])
     } else {
       // Fine. We need a worktree. But we don't have one. So let's create one.
       worktree = `${gitDir}/tmp-worktree`
       callGit(['--git-dir', gitDir, 'worktree', 'add', '--no-checkout', worktree])
       // No need for a full checkout, it's not as if we want to resolve any merge conflicts...
       callGit(['sparse-checkout', 'set'], worktree)
-      callGit(['switch', '-d', refName], worktree)
+      let head = callGit(['--git-dir', gitDir, 'rev-parse', 'HEAD'])
+      callGit(['switch', '-d', 'HEAD'], worktree)
       // Perform the merge
       callGit(['fetch', bundlePath, refName], worktree)
       callGit(['merge', '--no-edit', 'FETCH_HEAD'], worktree)
-      callGit(['update-ref', `refs/heads/${refName}`, 'HEAD'], worktree)
+      head = callGit(['rev-parse', 'HEAD'], worktree)
+      callGit(['--git-dir', gitDir, 'update-ref', 'HEAD', head])
     }
   }
 }
@@ -79,53 +81,65 @@ const getPushAuthorizationHeader = async (context, setSecret, appId, privateKey,
   return `Authorization: Basic ${auth}`
 }
 
-const pushRepositoryUpdate = async (context, setSecret, appId, privateKey, owner, repo, refName, bundlePath) => {
+const pushRepositoryUpdate = async (context, setSecret, appId, privateKey, owner, repo, refName, bundlePath, options) => {
+  options ||= {}
   context.log(`Pushing updates to ${owner}/${repo}`)
 
   // Updates to `build-extra` and `git-for-windows.github.io` need a worktree
   const bare = ['build-extra', 'git-for-windows.github.io'].includes(repo) ? '' : ['--bare']
-  const gitDir = `${repo}${bare ? '' : '/.git'}`
+  const branchOption = ['--single-branch', '--branch', refName, '--depth', '50']
+  const url = `https://github.com/${owner}/${repo}`
+  callGit(['clone', ...bare, ...branchOption, url, repo])
 
-  callGit(['clone', ...bare,
-    '--single-branch', '--branch', 'main', '--depth', '50',
-    `https://github.com/${owner}/${repo}`, repo
-  ])
+  const gitDir = `${repo}${bare ? '' : '/.git'}`
 
   if (bundlePath) {
     // Allow Git to fetch non-local objects by pretending to be a partial clone
     callGit(['--git-dir', gitDir, 'config', 'remote.origin.promisor', 'true'])
     callGit(['--git-dir', gitDir, 'config', 'remote.origin.partialCloneFilter', 'blob:none'])
-    mergeBundle(gitDir, !bare && repo, bundlePath, refName)
+    mergeBundle(gitDir, !bare && repo, bundlePath, options.extraPushRefs ? options.extraPushRefs[0] : refName)
   }
 
   if (repo === 'build-extra') {
     // Add `versions/package-versions-$ver*.txt`
     const fs = require('fs')
-    const ver = fs.readFileSync('bundle-artifacts/ver').toString().trim()
-    fs.renameSync(
-      'installer-x86_64/package-versions.txt',
-      `${repo}/versions/package-versions-${ver}.txt`
-    )
-    fs.renameSync(
-      'mingit-x86_64/package-versions.txt',
-      `${repo}/versions/package-versions-${ver}-MinGit.txt`
-    )
-    callGit([
-      'add',
-      `versions/package-versions-${ver}.txt`,
-      `versions/package-versions-${ver}-MinGit.txt`
-    ], repo)
-    callGit([
-      'commit',
-      '-s',
-      '-m', `versions: add v${ver}`,
-      `versions/package-versions-${ver}.txt`,
-      `versions/package-versions-${ver}-MinGit.txt`
-    ], repo)
+    if (fs.existsSync('bundle-artifacts/ver')) {
+      const filesToCommit = []
+      const ver = fs.readFileSync('bundle-artifacts/ver').toString().trim()
+      if (!options.mingitOnly) {
+        fs.renameSync(
+          'installer-x86_64/package-versions.txt',
+          `${repo}/versions/package-versions-${ver}.txt`
+        )
+        callGit([
+          'add',
+          `versions/package-versions-${ver}.txt`
+        ], repo)
+        filesToCommit.push(`versions/package-versions-${ver}.txt`)
+      }
+      fs.renameSync(
+        'mingit-x86_64/package-versions.txt',
+        `${repo}/versions/package-versions-${ver}-MinGit.txt`
+      )
+      callGit([
+        'add',
+        `versions/package-versions-${ver}-MinGit.txt`
+      ], repo)
+      filesToCommit.push(`versions/package-versions-${ver}-MinGit.txt`)
+      callGit([
+        'commit',
+        '-s',
+        '-m', `versions: add v${ver}`,
+        '--',
+        ...filesToCommit
+      ], repo)
+    }
 
-    // Update `download-stats.sh`
-    callProg('sh', ['./download-stats.sh', '--update'], repo)
-    callGit(['commit', '-s', '-m', 'download-stats: new Git for Windows version', './download-stats.sh'], repo)
+    if (!options.mingitOnly && owner === 'git-for-windows' && refName === 'main') {
+      // Update `download-stats.sh`
+      callProg('sh', ['./download-stats.sh', '--update'], repo)
+      callGit(['commit', '-s', '-m', 'download-stats: new Git for Windows version', './download-stats.sh'], repo)
+    }
   } else if (repo === 'git-for-windows.github.io') {
     callGit(['switch', '-C', 'main', 'origin/main'], repo)
     callProg('node', ['bump-version.js', '--auto'], repo)
@@ -135,7 +149,7 @@ const pushRepositoryUpdate = async (context, setSecret, appId, privateKey, owner
   const authorizationHeader = await getPushAuthorizationHeader(context, setSecret, appId, privateKey, owner, repo)
   callGit(['--git-dir', gitDir,
     '-c', `http.extraHeader=${authorizationHeader}`,
-    'push', `https://github.com/${owner}/${repo}`, refName
+    'push', `https://github.com/${owner}/${repo}`, refName, ...(options.extraPushRefs || [])
   ])
   context.log(`Done pushing ref ${refName} to ${owner}/${repo}`)
 }
